added listeners on subject service

+++ src/main/java/com/solvd/schedulegenerator/App.java
@@ -6,10 +6,10 @@ import com.solvd.schedulegenerator.menu.impl.ConsoleMenu;
 public class App {
     public static void main(String[] args) {
         IMenu menu = new ConsoleMenu();
-        int choice = 0;
+        Integer choice = 0;
         while (true) {
             menu.displayMenu();
-            choice = menu.getUserChoice();
+            choice = (Integer) menu.getUserChoice();
             menu.performUserChoice(choice);
         }
     }
